requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.2
graphql-core==3.2.3
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
strawberry-graphql==0.177.3
typing_extensions==4.5.0
uvicorn==0.22.0
